onpoint

Growing collection of small helper functions for point pattern analysis. Most functions are designed to work with the spatstat package.

See also

Author

Maintainer: Maximilian H.K. Hesselbarth mhk.hesselbarth@gmail.com (ORCID)