simulate_antecedent_conditions

Simulate heterogenous pattern

simulate_antecedent_conditions(x, i, j, nsim)

Arguments

x
ppp

i
mark of points that are randomized.

j
mark of points that do not change.

nsim
Number of patterns to simulate.

Details

Simulate point patterns as null model data for spatstat::envelope() using antecendet conditions as null model. x must be marked point pattern. Antecedent conditions are suitable as a null model if points of type j may influence points of type i, but not the other way around (Velazquez et al 2016). One example are the positions of seedlings that may be influenced by the position of mature trees.

References

Velazquez, E. et al. 2016. An evaluation of the state of spatial point pattern analysis in ecology. - Ecography (Cop.). 39: 1-14.

Wiegand, T. and Moloney, K. A. 2004. Rings, circles, and null models for point pattern analysis in ecology. - Oikos 104: 209-229.

See also

Examples

set.seed(42)
pattern_a <- spatstat::runifpoint(n = 20)
spatstat::marks(pattern_a) <- "a"
pattern_b <- spatstat::runifpoint(n = 100)
spatstat::marks(pattern_b) <- "b"
pattern <- spatstat::superimpose(pattern_a, pattern_b)

null_model <- simulate_antecedent_conditions(x = pattern, i = "b", j = "a", nsim = 19)
spatstat::envelope(Y = pattern, fun = spatstat::pcf, nsim = 19, simulate = null_model)

#> Extracting 19 point patterns from list  ...
#> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18,  19.
#> 
#> Done.

#> Pointwise critical envelopes for g(r)
#> and observed value for 'pattern'
#> Edge correction: "iso"
#> Obtained from 19 point pattern datasets in user-supplied list
#> Alternative: two.sided
#> Significance level of pointwise Monte Carlo test: 2/20 = 0.1
#> ......................................................................
#>       Math.label     Description                                      
#> r     r              distance argument r                              
#> obs   hat(g)[obs](r) observed value of g(r) for data pattern          
#> mmean bar(g)(r)      sample mean of g(r) from simulations             
#> lo    hat(g)[lo](r)  lower pointwise envelope of g(r) from simulations
#> hi    hat(g)[hi](r)  upper pointwise envelope of g(r) from simulations
#> ......................................................................
#> Default plot formula:  .~r
#> where "." stands for 'obs', 'mmean', 'hi', 'lo'
#> Columns 'lo' and 'hi' will be plotted as shading (by default)
#> Recommended range of argument r: [0, 0.25]
#> Available range of argument r: [0, 0.25]